requirements.txt
python-telegram-bot[async]>=20.0
pymongo==4.3.3
requests==2.27.1 
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
python-dotenv==1.0.1
typing_extensions==4.12.0
yarl==1.9.4
geopy==2.2.0
SQLAlchemy[asyncio]>=2.0
asyncpg
psycopg2-binary
nest-asyncio